Bugfix: Don't crash if kill called after process closed (#25)
* Add failing test

* Get test green on Mac

* Formatting

+++ src/ChildProcess.re
@@ -146,14 +146,15 @@ let _spawn =
 
   let retStdin: inputPipe = {write: stdinWrite, close: stdinClose};
 
-  let kill = sig_ => {
-    let signalToUse =
-      Sys.win32
-        ? Sys.sigkill  /* Sigkill is the only signal supported on Win by the Unix module */
-        : sig_;
+  let kill = sig_ =>
+    if (isRunning^) {
+      let signalToUse =
+        Sys.win32
+          ? Sys.sigkill  /* Sigkill is the only signal supported on Win by the Unix module */
+          : sig_;
 
-    Unix.kill(pid, signalToUse);
-  };
+      Unix.kill(pid, signalToUse);
+    };
 
   let ret: innerProcess = {
     pid,
